backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pytesseract==0.3.10
Pillow==10.3.0
opencv-python-headless==4.9.0.80
pdfplumber==0.11.0
pdf2image==1.17.0
fpdf2==2.7.9
python-magic==0.4.27
slowapi==0.1.9
